Add some check to run.sh

+++ benchmarks/run.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 DIR=/tmp/fio
-mkdir $DIR
+if [ ! $# -eq 1 ]; then
+	echo "Usage: $0 <benchmark>.fio"
+	exit 1
+fi
+mkdir -p $DIR
 fio $1
 rm -rf $DIR
